designs/utils/image_recognition.py
import os
from pathlib import Path
import numpy as np
from PIL import Image

class DesignImageRecognition:
    @staticmethod
    def preprocess_image(image_path):
        """Preprocess image for recognition model"""
        try:
            img = Image.open(image_path)
            # Convert to RGB if necessary
            if img.mode != 'RGB':
                img = img.convert('RGB')
            # Resize to standard size
            img = img.resize((224, 224))
            # Convert to numpy array and normalize
            img_array = np.array(img) / 255.0
            return img_array
        except Exception as e:
            print(f"Error preprocessing image: {e}")
            return None

    @staticmethod
    def analyze_design(image_path):
        """Analyze design image for pattern recognition"""
        try:
            processed_img = DesignImageRecognition.preprocess_image(image_path)
            if processed_img is None:
                return None
            
            # Placeholder for actual model prediction
            # Will implement actual model inference here
            results = {
                'pattern_type': 'geometric',
                'dominant_colors': ['#ffffff', '#000000'],
                'complexity_score': 0.75,
                'similarity_score': 0.0
            }
            return results
        except Exception as e:
            print(f"Error analyzing design: {e}")
            return None

    @staticmethod
    def extract_features(image_path):
        """Extract features from design image"""
        try:
            processed_img = DesignImageRecognition.preprocess_image(image_path)
            if processed_img is None:
                return None
            
            # Placeholder for feature extraction
            # Will implement actual feature extraction here
            features = {
                'color_histogram': [],
                'edge_features': [],
                'texture_features': []
            }
            return features
        except Exception as e:
            print(f"Error extracting features: {e}")
            return None