requirements.txt
Django>=5.1.7
Pillow>=10.0.0
numpy>=1.24.0
scikit-learn>=1.4.0
scikit-image>=0.22.0
opencv-python>=4.8.0
torch>=2.2.0
torchvision>=0.17.0